+++ website/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    leptos_bulma::LeptosBulma::build("./style");
+}