+++ website/src/components.rs
@@ -1,0 +1,7 @@
+use leptos::*;
+use leptos_meta::Title;
+
+#[component]
+pub fn PageTitle(#[prop(into)] text: TextProp) -> impl IntoView {
+    view! { <Title text=format!("{} | Leptos Bulma", text.get())/> }
+}