+++ website/src/main.rs
@@ -1,0 +1,9 @@
+#[cfg(feature = "ssr")]
+fn main() {
+    // Unused; required for cargo-leptos to build the front as it also wants to build a bin
+}
+
+#[cfg(not(feature = "ssr"))]
+fn main() {
+    // Unused
+}