+++ website/src/pages/mod.rs
@@ -1,0 +1,5 @@
+mod home_page;
+mod not_found_page;
+
+pub use home_page::HomePage;
+pub use not_found_page::NotFoundPage;