+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- A Leptos application to replace the mdBook.
+
 ## [0.0.0-alpha.7] - 2024-03-27
 
 ### Added